Improve ConvNet using the CaloGAN paper.
Make the convolutional model much more similar in architecture to the
CaloGAN model, which uses LAGAN-style blocks in three streams, just as
described in the original paper. Also, tweak certain training
parameters and optimize transferring tensors to GPU. Significant
improvements have been achieved in the performance of the ConvNet model.